% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_error_functions.R
\name{internal_functions_II}
\alias{internal_functions_II}
\alias{exp_error_simple}
\alias{exp_error_multiple}
\alias{exp_error_convergence}
\title{Internal Functions II}
\usage{
exp_error_simple(x, y, weights_x = NULL, weights_y = NULL,
  sd_x = NULL, sd_y = NULL, r_xy_errors = NULL)

exp_error_multiple(x, y, weights_x = NULL, weights_y = NULL,
  sd_x = NULL, sd_y = NULL, r_xy_errors = NULL,
  approx_solution = FALSE)

exp_error_convergence(count, max_iterations, slope_per_iteration)
}
\description{
Functions used in york function. Internal Functions II
 compromise all possible error messages.

 The exp_error_simple function returns all possible error messages
 if the input is a simple regression with only one sample of x and y.

 The exp_error_multiple function returns all possible error messages if
 mult.sample = TRUE is specified

 The exp_error_convergence function returns an error message if the slope
 coefficient does not converge. It will also print the value if the last
 determined slope coefficients.
}
\keyword{internal}
