% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_plot_internal.R
\name{method_plot_internal}
\alias{method_plot_internal}
\alias{f_p_influential}
\alias{f_p_chisq_test}
\title{Internal Functions for Plot Method}
\usage{
f_p_influential(x_data, y_data, mult_samples, sd_x, sd_y, r_xy_errors,
  coef, coef_se, tolerance, max_iterations, x_mult, y_mult)

f_p_chisq_test(chisq_df, p_value, chisq_statistic)
}
\description{
Functions used in plot.york method. For functions in "Internal Functions
 for Plot Method" the output is used for proceeding calculations or
 displaying in the plot.york method. All the functions in "Internal
 Functions for Plot Method" begin with f_p_ and then the name of the
 function is given.

 f_p_influential determines influential points in a york regression.
 An influential point is defined as a point that significantly changes
 the slope coefficient if omitted. Assymptotic normality of the slope
 coefficient is assumed. The test is based on a significance level
 of alpha = 0.01.

 f_p_chisq_test is a function that calculates the critical values for
 alpha = c(0.1, 0.05, 0.01) and the corresponding probabilities in a
 \eqn{\chi^2}--test. Additionally, values for the rejection region
 for alpha = 0.01 are given.
}
\keyword{internal}
