% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_predict.R
\name{predict.york}
\alias{predict.york}
\title{Predict method for York Model Fits}
\usage{
\method{predict}{york}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a fitted object of class inheriting from "york".}

\item{newdata}{a vector or a dataframe to be used for the prediction.
If not specified the original x-values with their respective
fitted y-values will be returned.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
Returns a list, containing a dataframe with predicted values.

\describe{

  \item{prediction}{a dataframe containing either the original x-values
    or the specified newdata in the first column and the corresponding
    predicted values in the second column.}
}
}
\description{
Predictions from a model fitted by \code{\link{york}}.
}
\examples{
 # Example: York's regression with weight data taken from Pearson (1901):
 x <- c(0.0, 0.9, 1.8, 2.6, 3.3, 4.4, 5.2, 6.1, 6.5, 7.4)
 y <- c(5.9, 5.4, 4.4, 4.6, 3.5, 3.7, 2.8, 2.8, 2.4, 1.5)
 weights_x <- c(1e+3, 1e+3, 5e+2, 8e+2, 2e+2, 8e+1, 6e+1, 2e+1, 1.8, 1)
 weights_y <- c(1, 1.8, 4, 8, 20, 20, 70, 70, 1e+2, 5e+2)

 # fit york model
 york_fit <- york(x, y, weights_x, weights_y, r_xy_errors = 0)

 # values for prediction
 new <- c(2,3,4,3.6)
 predict(york_fit, new)

}
