% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york_plots.R
\name{york.plots}
\alias{york.plots}
\title{Plots for York's regression}
\usage{
york.plots(york.output)
}
\arguments{
\item{york.output}{An object of class "york"}
}
\value{
york.plots returns 6 different diagnostic plots.
}
\description{
Gives several diagnostic plots for York's regression.
}
\details{
Given an object of class "york" this function gives several
dianostic plots. The first plot shows York's best-fit straight line only. The
second plot shows ork' best fit straight line compared to OLS and orthogonal
regression. Additionally, the different "center of gravity" are shown. The
third plot is a "y-residuals vs. x-residuals plot". The fourth is a
"x-residuals vs. fitted-y plot". The fith plot is a "y-residuals vs.
fitted y" plot and the last plot is a "trace plot" which shows the slope after
each interation, i.e. how the slope coefficient converges.
}
\examples{
# Example: York's regression with weight data taken from Pearson (1901):
x <- c(0.0, 0.9, 1.8, 2.6, 3.3, 4.4, 5.2, 6.1, 6.5, 7.4)
y <- c(5.9, 5.4, 4.4, 4.6, 3.5, 3.7, 2.8, 2.8, 2.4, 1.5)
weights.x <- c(1e+3, 1e+3, 5e+2, 8e+2, 2e+2, 8e+1, 6e+1, 2e+1, 1.8, 1)
weights.y <- c(1, 1.8, 4, 8, 20, 20, 70, 70, 1e+2, 5e+2)
r.xy <- 0
york.output <- york(x, y, weights.x = weights.x, weights.y = weights.y,
                    r.xy = 0, mult.samples = T, exact.solution = FALSE)))
york.plots(york.output)
}
