% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york_predict.R
\name{york.predict}
\alias{york.predict}
\title{Predict Method for York Fits}
\usage{
york.predict(york.output, newdata)
}
\arguments{
\item{york.output}{A fitted object of class inheriting from "york".}

\item{newdata}{A vector or a dataframe in which to look for variables
with which to predict.}
}
\value{
Returns a list, containing a dataframe with predicted values
as first element.

  \describe{

  \item{prediction}{A dataframe containing the \code{newdata} in the first
  column and the corresponding predicted values in the second column.}}
}
\description{
Obtains predictions from a fitted York's regression.
}
\examples{
# Example: York's regression with weight data taken from Pearson (1901):
x <- c(0.0, 0.9, 1.8, 2.6, 3.3, 4.4, 5.2, 6.1, 6.5, 7.4)
y <- c(5.9, 5.4, 4.4, 4.6, 3.5, 3.7, 2.8, 2.8, 2.4, 1.5)
weights.x <- c(1e+3, 1e+3, 5e+2, 8e+2, 2e+2, 8e+1, 6e+1, 2e+1, 1.8, 1)
weights.y <- c(1, 1.8, 4, 8, 20, 20, 70, 70, 1e+2, 5e+2)
fit <- york(x = x, y = y, weights.x = weights.x, weights.y = weights.y)

# predict with:
data<- c(2,3,4,3.6)
york.predict(york.output = fit, newdata = data)

}
