% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithm_with_documentation.R
\name{york}
\alias{york}
\title{Simple linear regression of X and Y-variables with correlated errors.}
\arguments{
\item{x}{A 1 times n numeric row vector.of the \code{X} Variable}

\item{y}{A 1 times n numeric row vector.of the \code{Y} Variable}

\item{tolerance}{The tolerance for convergence which is set a priori to
\code{1e-10}}

\item{weights.x}{The prespecified 1 times n weights vector for x values}

\item{weights.y}{The prespecified 1 times n weights vector for y values}

\item{r.xy}{The prespecified correlation coefficient between X and Y}

\item{sd.x}{The standard error of the x values}

\item{sd.y}{The standard error of the y values}
}
\value{
yorks Returns an object of class "york" the York Regression for the
\code{x} and \code{y} data for either specified weights \code{weights.x}
and \code{weights.y} or specified standard errors \code{sd.x} abd \code{sd.y}
An object of class "york" containing the following components:
coefficients.york (a matrix which contains the york estimates for intercept
and slope with the respective standard errors), coefficients.mayor
(a matrix which contains the deming estimates for intercept and slope with
the respective standard errors), coefficients.ols (a matrix which contains
the ols estimates for intercept and slope with the respective
standard errors), weighting.vector (the speciefied or calculated weights),
x.residuals (the York x-residuals), y.residuals (the York y-residuals),
fitted.y (the fitted York y-values), df.regression (the number of degress
of freedom of the York regression), mean.x (the ordinary mean of X), mean.y
(the ordinary mean of Y), reduced.chisq (the reduced chi-squared statistic
i.e. the goodness of fit measure of York's regression), std.Error.chisq
(the standard error of the chi-squared statistic), number.of.iterations
(the total number of iterations), slope.after.each.iteration (the york slope
after each interation), original.x.values (the original x values),
original.y.values (the original y values), fitted.y.ols (the fitted values
for OLS), se.of.reg.ols (the standard error of the regression for OLS),
fitted.y.orthogonal (the fitted values for deming regression), data
(a data matrix which contains as colums the x-, y-,sd.x- and sd.y-values)
}
\description{
Implements the algorithm for the problem of the best-ﬁt straight line to
independent points with errors in both x and y general York 1969 solution according to
the algorithm of Wehr & Saleska (2017)
}
\details{
Given \eqn{n} pairs of \eqn{(X_i, Y_i), i = 1, \ldots, n}, their weights
\eqn{(\omega(X_i), \omega(Y_i), i = 1, \ldots, n} or their standard errors
\eqn{SE(X_i)} and \eqn{SE(Y_i)}, the \code{york} function finds the best fit
straight line using the algorithm of York et al. (1966)/ York et al. (1969)
as presented in Wehr & Saleska (2017). If the data contains NA values then
the share of NA values in the total values is calculated und the rows with NA
 values will be deleted.
}
\examples{
york(x, y, tolerance = 1e-10, weights.x = NULL, weights.y = NULL,
r.xy = NULL, sd.x = NULL, sd.y = NULL)
\dontrun{
# Example: No standard errors or weights specified
york(x, y, r.xy = 0)
# Example: You can't specify weights and standard errors at the same time
york(x , y, sd.x, sd.y, weights.x, weights.y, r.xy = 0)
# Example: x and y must have same length
york(x = c(0.0, 0.9, 1.8, 2.6, 3.3, 4.4, 5.2, 6.1, 6.5),
y = c(5.9, 5.4, 4.4, 4.6, 3.5, 3.7, 2.8, 2.8, 2.4, 1.5),
weights.x = weights.x, weights.y = weights.y, r.xy = 0)
}
}
