% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{package}
\name{yorkregression}
\alias{yorkregression}
\alias{yorkregression-package}
\title{yorkregression: A package to fit a linear York's model.}
\description{
The yorkregression package provides one main function named york and three methods
summary.york, plot.york and predict.york.
}
\section{york function}{

 The york function fits a York model to the data. It is used when
 both X and Y variables are subject to measurement error. Returned
 is an object of class york.
}

\section{york methods}{

 The summary.york method returns the summary statistics of a fitted york
 model.

 The plot.york method returns plots for the fitted york model.

 The predict.york method returns predicted values for y for a fitted york
 model.
}

